Add support for `offset` in detecting if `start` and `end` should switch
Currently, an end-container always has to be after a
start-container: a range starting at the first word of a
sentence (`range.setStart(sentence.head)`) cannot end at a
relative reference to its last node
(`range.setEnd(sentence, Infinity)`).

This addition looks at `offset` during run-time to detect which
node is references, and if that node is before or after the new
node.

Closes GH-1.